Started move of gerty api calls into helpers.py
