+++ lnbits/extensions/gerty/__init__.py
@@ -9,9 +9,7 @@ from lnbits.tasks import catch_everything_and_restart
 
 db = Database("ext_gerty")
 
-
 gerty_ext: APIRouter = APIRouter(prefix="/gerty", tags=["Gerty"])
-
 
 def gerty_renderer():
     return template_renderer(["lnbits/extensions/gerty/templates"])
